test: Fix core tests
