+++ .env.example
@@ -1,0 +1,7 @@
+# Example of an environment file for the CMIP-REF project
+# This allows for running the project in a development environment outside of a container
+
+CELERY_BROKER_URL=redis://localhost:6379/1
+
+REF_OUTPUT_ROOT=out
+REF_ESGF_ROOT=.esgpull/data
