+++ conftest.py
@@ -15,3 +15,20 @@ def esgf_data_dir() -> Path:
     pull = esgpull.Esgpull()
 
     return pull.config.paths.data
+
+
+@pytest.fixture
+def test_dataset(esgf_data_dir) -> Path:
+    return (
+        esgf_data_dir
+        / "CMIP6"
+        / "ScenarioMIP"
+        / "CSIRO"
+        / "ACCESS-ESM1-5"
+        / "ssp126"
+        / "r1i1p1f1"
+        / "Amon"
+        / "tas"
+        / "gn"
+        / "v20210318"
+    )
