+++ ruff.toml
@@ -29,7 +29,8 @@ ignore = [
 "test*.py" = [
     "D",  # Documentation not needed in tests
     "S101",  # S101 Use of `assert` detected
-    "PLR2004" # Magic value used in comparison
+    "PLR2004", # Magic value used in comparison
+    "PLR0913", # Too many arguments in function definition
 ]
 "conftest.py" = [
     "D",  # Documentation not needed in tests
